requirements.txt
flask
flask_cors
flask_restful
flask_sqlalchemy
Flask-Security-Too
email_validator
bcrypt